Removed the Parse function from the Formatter interface, and in the process cleaned up a lot of code
