+++ formatters/clover/clover.go
@@ -14,8 +14,7 @@ import (
 var searchPaths = []string{"build/logs/clover.xml", "clover.xml"}
 
 type Formatter struct {
-	Path        string
-	SourceFiles []formatters.SourceFile
+	Path string
 }
 
 func (f *Formatter) Search(paths ...string) (string, error) {
@@ -31,16 +30,21 @@ func (f *Formatter) Search(paths ...string) (string, error) {
 	return "", errors.WithStack(errors.Errorf("could not find any files in search paths for clover. search paths were: %s", strings.Join(paths, ", ")))
 }
 
-func (f *Formatter) Parse() error {
-	fx, err := os.Open(f.Path)
+func (r Formatter) Format() (formatters.Report, error) {
+	rep, err := formatters.NewReport()
 	if err != nil {
-		return errors.WithStack(err)
+		return rep, err
+	}
+
+	fx, err := os.Open(r.Path)
+	if err != nil {
+		return rep, errors.WithStack(err)
 	}
 
 	c := &xmlFile{}
 	err = xml.NewDecoder(fx).Decode(c)
 	if err != nil {
-		return errors.WithStack(err)
+		return rep, errors.WithStack(err)
 	}
 
 	gitHead, _ := env.GetHead()
@@ -49,7 +53,7 @@ func (f *Formatter) Parse() error {
 			num := 1
 			sf, err := formatters.NewSourceFile(pf.Name, gitHead)
 			if err != nil {
-				return errors.WithStack(err)
+				return rep, errors.WithStack(err)
 			}
 			for _, l := range pf.Lines {
 				for num < l.Num {
@@ -60,22 +64,11 @@ func (f *Formatter) Parse() error {
 				sf.Coverage = append(sf.Coverage, ni)
 				num++
 			}
-			sf.CalcLineCounts()
-			f.SourceFiles = append(f.SourceFiles, sf)
+			err = rep.AddSourceFile(sf)
+			if err != nil {
+				return rep, errors.WithStack(err)
+			}
 		}
-	}
-
-	return nil
-}
-
-func (r Formatter) Format() (formatters.Report, error) {
-	rep, err := formatters.NewReport()
-	if err != nil {
-		return rep, err
-	}
-
-	for _, f := range r.SourceFiles {
-		rep.AddSourceFile(f)
 	}
 
 	return rep, nil
