+++ formatters/simplecov/simplecov.go
@@ -3,9 +3,7 @@ package simplecov
 import (
 	"encoding/json"
 	"os"
-	"sort"
 	"strings"
-	"time"
 
 	"github.com/Sirupsen/logrus"
 	"github.com/codeclimate/test-reporter/env"
@@ -16,8 +14,7 @@ import (
 var searchPaths = []string{"coverage/.resultset.json"}
 
 type Formatter struct {
-	Path  string
-	Tests []Test
+	Path string
 }
 
 func (f *Formatter) Search(paths ...string) (string, error) {
@@ -33,60 +30,41 @@ func (f *Formatter) Search(paths ...string) (string, error) {
 	return "", errors.WithStack(errors.Errorf("could not find any files in search paths for simplecov. search paths were: %s", strings.Join(paths, ", ")))
 }
 
-func (f *Formatter) Parse() error {
-	if f.Path == "" {
-		_, err := f.Search()
-		if err != nil {
-			return errors.WithStack(err)
-		}
-	}
-	jf, err := os.Open(f.Path)
+func (r Formatter) Format() (formatters.Report, error) {
+	rep, err := formatters.NewReport()
 	if err != nil {
-		return errors.WithStack(err)
+		return rep, err
 	}
+
+	jf, err := os.Open(r.Path)
+	if err != nil {
+		return rep, errors.WithStack(err)
+	}
+
 	m := map[string]input{}
 	err = json.NewDecoder(jf).Decode(&m)
 	if err != nil {
-		return errors.WithStack(err)
+		return rep, errors.WithStack(err)
 	}
-	f.Tests = make([]Test, 0, len(m))
-	for k, v := range m {
-		tt := Test{
-			Name:        k,
-			Timestamp:   v.Timestamp.Time(),
-			SourceFiles: make([]formatters.SourceFile, 0, len(v.Coverage)),
-		}
+
+	for _, v := range m {
 		gitHead, _ := env.GetHead()
 		for n, ls := range v.Coverage {
 			fe, err := formatters.NewSourceFile(n, gitHead)
 			if err != nil {
-				return errors.WithStack(err)
+				return rep, errors.WithStack(err)
 			}
 			fe.Coverage = ls
-			fe.CalcLineCounts()
-			tt.SourceFiles = append(tt.SourceFiles, fe)
+			err = rep.AddSourceFile(fe)
+			if err != nil {
+				return rep, errors.WithStack(err)
+			}
 		}
-		sort.Slice(tt.SourceFiles, func(a, b int) bool {
-			return tt.SourceFiles[a].Name < tt.SourceFiles[b].Name
-		})
-		f.Tests = append(f.Tests, tt)
 	}
-	return nil
-}
 
-type Test struct {
-	Name        string
-	Timestamp   time.Time
-	SourceFiles []formatters.SourceFile
-}
-
-type rubyTime int64
-
-func (rt rubyTime) Time() time.Time {
-	return time.Unix(int64(rt), 0)
+	return rep, nil
 }
 
 type input struct {
-	Timestamp rubyTime                       `json:"timestamp"`
-	Coverage  map[string]formatters.Coverage `json:"coverage"`
+	Coverage map[string]formatters.Coverage `json:"coverage"`
 }
